Create TempAVL.java
Looked over everything... its pretty tricky to pinpoint what the issue is exactly but I believe its coming from the left and right rotations for the reason that we arent changing what 'this.root' points to after a L/R rotation... I think everything else is okay though. An example of how 'this.root' is having issues is if we have a tree with 10, 16 and we insert 30... well, that means we must perform a left rotation around 16 and then set this.root to equal 16 instead of 10, but that didn't seem like the case with what we have thus far. I wasn't able to fix everything but the code I filled in to leftRotation() and rightRotation() should help you understand where Im coming from.